split logic to hook
