+++ apps/medon-fe/src/components/AvailabilityCalendar/style.tsx
