+++ apps/medon-fe/src/translation/en/translation.json
@@ -323,6 +323,7 @@
   },
   "availability": {
     "title": "Availability schedule:",
-    "message": "Select day or availability slot to start editing"
+    "message": "Select day or availability slot to start editing",
+    "timeUsed": "That time is already used"
   }
 }
